Optimization for checking if a file exists.
Before it was doing a file_get_contents on the entire file whether or
not the data has changed, now it only does it if it's going to go
through the effort of doing a PUT request. This takes it down from
around 30m to around 30s for a second run of the script with no changes.